% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5.R
\docType{data}
\name{big5}
\alias{big5}
\title{Big 5 Personality Factors Survey Data}
\format{A dataframe}
\source{
\href{http://personality-testing.info/_rawdata/}{Open psychology data: Raw data from online personality tests}
}
\usage{
data(big5)
}
\description{
Answers to the Big Five Personality Test, constructed with items from the International Personality Item Pool.
}
\examples{
data(big5)

}
\keyword{datasets}
