% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequencies}
\usage{
freq(df, fn = NULL, maxrow = 30, trim = TRUE, type = "html",
  template = NULL)
}
\arguments{
\item{df}{Input dataframe.}

\item{fn}{File name. Optional file name if wanting to save the output.}

\item{maxrow}{Maximum number of rows to display in each frequency table.}

\item{trim}{Trim whitespace of character vectors.}

\item{type}{Output type. Either html or doc.}

\item{template}{Word template. Optional doc template to use if producing doc output.}
}
\value{
A frequency table in html or doc format.
}
\description{
This function generates frequencies
}
\examples{

\dontrun{
# Create frequency tables for the entire dataset
freq(big5)

# For specific variable/s
freq(big5$race)
freq(big5[4:6])

# To automatically open html output in your browser use the following option:
options(frequencies_open_output = TRUE)
freq(big5[, c('gender', 'E1')])

# To save the output specify the filename and format
freq(big5, fn = "mydir/myfile.html")

# Produce a list of tables and flextables
out <- freq(big5)
out$tables[1] # standard output to console
out$flextables[4] # flextable output to Viewer in RStudio

#Suppress Viewer output
options(frequencies_output_flextables = FALSE)
x <- freq(big5)
x[5] # standard print output

}

}

