% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\docType{package}
\name{frequencies}
\alias{frequencies}
\alias{frequencies-package}
\title{frequencies: A package to generate frequencies}
\description{
The frequencies package comprises functions to generate SPSS like frequencies in R.
}
\section{Frequency functions}{


\code{\link{freq}}
}

