% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\docType{package}
\name{frequency}
\alias{frequency}
\alias{frequency-package}
\title{frequency: A package to generate frequencies}
\description{
The frequency package comprises functions to generate SPSS like frequencies in R.
}
\section{Frequency functions}{


\code{\link{freq}}
}

