% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{print.freq_table}
\alias{print.freq_table}
\title{Print frequency tables}
\usage{
\method{print}{freq_table}(x, ...)
}
\arguments{
\item{x}{object of class \code{freq.table}}

\item{...}{optional arguments to \code{data.frame}}
}
\description{
S3 method for class 'freq_table'
}
\examples{
# Suppress external output for examples
options(frequency_render = FALSE)

x <- freq(big5[1])
print(x)
print(x[[1]])

}
