% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequencies}
\usage{
freq(df, fn = NULL, maxrow = 30, trim = TRUE, type = "html",
  template = NULL)
}
\arguments{
\item{df}{Input dataframe.}

\item{fn}{File name. Optional file name if wanting to save the output.}

\item{maxrow}{Maximum number of rows to display in each frequency table.}

\item{trim}{Trim whitespace of character vectors.}

\item{type}{Output type. Either html or doc.}

\item{template}{Word template. Optional doc template to use if producing doc output.}
}
\value{
A frequency table in html or doc format.
}
\description{
This function generates frequencies
}
\examples{
tmp <- c(1:3)
freq(tmp)

}

