% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency-package.R
\docType{package}
\name{frequency}
\alias{frequency}
\alias{frequency-package}
\title{frequency: A package to generate frequencies}
\description{
The frequency package comprises functions to generate SPSS like frequencies in R.
}
\section{Frequency functions}{


\code{\link{freq}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilcoxa/frequency}
  \item Report bugs at \url{https://github.com/wilcoxa/frequency/issues}
}

}
\keyword{internal}
